feat: Add Tag Api - GET /tag/artist/{tag_id}

+++ backend/src/main/java/com/lemonmul/planetdhance/repo/TagRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TagRepo extends JpaRepository<Tag,Integer> {
+public interface TagRepo extends JpaRepository<Tag,Long> {
 
     //해당 검색어 포함한 태그 리스트 (검색 빈도 순)
     Slice<Tag> findByNameContainingOrderByHitDesc(String searchStr, Pageable pageable);
